Switch back to using `main` branch of `per1234/workflowsawaiting`
Some time ago, I started experienced mysterious spurious failures of the workflow run due to errors during GitHub API
requests made via the workflowsawaiting script.

I was not able to determine the cause of these errors so I implemented what was intended to be a provisional workaround
in the script. I pushed that to a temporary `dev` branch in the repository and configured the workflow to use that
version of the script.

I had hoped this might have been caused by a transient problem in the GitHub infrastructure that would eventually be
fixed, after which I could discard the workarounds from the script. However, I find it still occurs.

Staging the workaround in a separate branch in the script repo is not sustainable for the long term, so I have now
merged those changes into the `main` branch of the script repo and deleted the superfluous `dev` branch. The workflow is
hereby updated to use the version of the script from the `main` branch of the repo.