Manage Python version used to run script via "Find Workflows Awaiting Approval" workflow
Previously, the workflow used whichever version of Python was provided as default by the "actions/setup-python" action.
That version can change at any time, which might result in instability.

The version of Python used by the workflow is now controlled by the version specified in the pyproject.toml file of the
workflowsawaiting project.