Replace unsupported flag in script dependencies installation command
workflowsawaiting has several Python package dependencies which must be installed in order to use the script. The
dependencies are managed using the Poetry tool.

In addition to the script dependencies, various Python packages are used in the infrastructure of the workflowsawaiting
project (development dependencies), so those are also specified as dependencies of the project. For the sake of
efficiency, the dependencies installation command in the "Find Workflows Awaiting Approval" workflow is configured to
only install the script dependencies. This is possible because the development dependencies are organized into a
dedicated "dev" dependencies group.

Previously, the `--no-dev` flag was used in the `poetry install` command to accomplish this. However, Poetry has
migrated to a more flexible system where dependencies can be organized into any number of arbitrary groups, and the
installation of the groups controlled. For this reason the "dev" group-specific `--no-dev` flag has been replaced by
general purpose flags that can be used with any group.

The previous dependencies installation command now fails when used with modern versions of Poetry:

```
The option "--no-dev" does not exist
```

The flag is hereby replaced with the modern equivalent `--only main` (the "main" group is the script dependencies).